requirements.txt
crewai
python-dotenv
langchain-openai
langchain-anthropic
langchain-community